Preserve scheduler state across runs and don't stop existing tasks.
Summary:
Currently, restarts of the scheduler cancels all existing tasks mesos
is running.
This records the scheduler state in a file on sigterm and restores
the state on start up.

Test Plan:
Mainly, try restarting the scheduler in staging while builds on happening.
Also try starting it after SIGINT and SIGKILL.
I added a test for the save and restore but it assumes there's a `changes_scheduler` module so it doesn't actually work now.
I plan to move the class into `changes_scheduler.py` in a later diff.

Reviewers: anupc, kylec

Reviewed By: kylec

Subscribers: nate, jukka, kylec

Projects: #changes-stability

Differential Revision: https://tails.corp.dropbox.com/D137628