Add statsreporter and use it trivially
Summary:
Not a meaningful use, but a useful mechanism and with just enough use to verify it is working.
statsreporter was copied with minimal tweaks from Changes, which is sad, but we don't yet have
a code sharing mechanism that seems preferable to simple copying here.

Test Plan: Light unit; intent to try in staging prior to push.

Reviewers: paulruan

Reviewed By: paulruan

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D155609