Interact with Changes via ChangesAPI
Summary:
For clarity and ease of testing, this moves us to using a ChangesAPI instance rather than hand-crafted urllib2
calls.
I mostly tried to preserve behavior, but there are probably some subtle differences worth considering.
It's somewhat of a problem that any leaked exceptions appear to kill the Scheduler and may not result in Sentry
errors.

Test Plan: Light unit

Reviewers: nate

Reviewed By: nate

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D156009