Change failover_timeout to 1 week as recommended in the Mesos docs
Summary:
See here: http://mesos.apache.org/documentation/latest/high-availability-framework-guide/ (under "Framework Reregistration").
We could probably survive with not quite that long, but in particular because
we save the framework id on shutdown, if the failover timeout is small and
we reach it, the master will tell us to go away if we come back up with the
same id. (Then we have to fully stop the scheduler, remove state.json, and then
restart it).

Test Plan: None

Reviewers: kylec, paulruan

Reviewed By: paulruan

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D172497