Prefer slaves which are likely to contain a snapshot to run jobsteps.
Summary:
Make a note of snapshot id when scheduling a jobstep on a slave.
If a jobstep using the same snapshot needs to be scheduled in the future, attempt to prioritize
sending it to the same slave(s) to increase chances of caching the snapshot.

Test Plan: Unit tests

Reviewers: kylec, nate

Reviewed By: nate

Subscribers: changesbot, paulruan

Maniphest Tasks: T82707

Differential Revision: https://tails.corp.dropbox.com/D177305