Clear the offers cache when the scheduler is disconnected from the master.
Summary:
When the Mesos master disappears, the scheduler is left with a collection of
stale offers. Clear all slaves and offers from the cache when a master
disconnect signal is received. Don't bother declining the offers, since the
master is already gone. A new master will provide a set of new, fresh
offers when it comes online.

Test Plan: Added unit test.

Reviewers: nate

Reviewed By: nate

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D188559